File Hashing

+++ auto_update.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-
 import sys
 import os
 import re
@@ -7,6 +6,7 @@ import json
 import requests
 import urllib.request
 import tarfile
+import hashlib
 
 try:
     installation = str(sys.argv[1])
@@ -14,7 +14,8 @@ except:
     installation = '/.steam/root/steamapps/compatdata/284160/pfx/drive_c/users/steamuser/AppData/Local/BeamNG.drive/'
 
 tags = []
-default = ['etk.tar.gz', 'jri.tar.gz', 'utah.tar.gz', 'derby.tar.gz', 'hirochi.tar.gz', 'industrial.tar.gz', 'gridmap_v2.tar.gz', 'small_island.tar.gz',]
+hash_table = []
+default = ['etk', 'jri', 'utah', 'derby', 'hirochi', 'industrial', 'gridmap_v2', 'small_island',]
 
 path = os.path.expanduser("~") + installation
 
@@ -43,27 +44,59 @@ for tag in response:
 print('\n\nStarting to update.\nYou can always cancel the download by pressing CTRL + C')
 print('\nLooking for modded maps...')
 
+#Generate Hashes for existing .dds
+def md5(fname):
+    hash_md5 = hashlib.md5()
+    with open(fname, "rb") as f:
+        for chunk in iter(lambda: f.read(4096), b""):
+            hash_md5.update(chunk)
+    return hash_md5.hexdigest()
+
+for dds in os.listdir(cache + 'terrainMaterialCache/'):
+    hash_table.append(md5(cache + 'terrainMaterialCache/' + dds) + '  ' + dds)
+
 try:
     for mod in repodb['mods']:
         mod_id = str(repodb['mods'][mod]['modData'].get('resource_id'))
         mod_title = str(repodb['mods'][mod]['modData'].get('title'))
         if mod_id in tags:
-            print('Getting textures for ' + mod_title)
-            url = 'https://github.com/SnoutBug/BeamNG_terrainMaterialCache/releases/download/' + mod_id + '/main.tar.gz'
-            filename = mod_id + '.tar.gz'
-            urllib.request.urlretrieve(url, filename)
-            with tarfile.open(filename) as tar:
-                tar.extractall(cache)
-            os.remove(filename)
+            print('Verifying ' + mod_title)
+            hash_url = 'https://github.com/SnoutBug/BeamNG_terrainMaterialCache/releases/download/' + mod_id + '/md5.txt'
+            hash_filename = mod_id + '_md5.txt'
+            urllib.request.urlretrieve(hash_url, hash_filename)
+            with open(hash_filename) as hash_file:
+                hashes = hash_file.read().splitlines()
+            os.remove(hash_filename)
+            for hash in hashes:
+                if not hash in hash_table:
+                    print('  > Downloading Files...')
+                    url = 'https://github.com/SnoutBug/BeamNG_terrainMaterialCache/releases/download/' + mod_id + '/main.tar.gz'
+                    filename = mod_id + '.tar.gz'
+                    urllib.request.urlretrieve(url, filename)
+                    with tarfile.open(filename) as tar:
+                        tar.extractall(cache)
+                    os.remove(filename)
+                    break #continue with next mod
 
     print('\nLooking for default maps...')
     url = 'https://github.com/SnoutBug/BeamNG_terrainMaterialCache/releases/download/default/'
     for filename in default:
-        print('Getting textures for ' + re.sub('\.tar.gz$', '', filename))
-        urllib.request.urlretrieve(url + filename, filename)
-        with tarfile.open(filename) as tar:
-            tar.extractall(cache)
-        os.remove(filename)
+        print('Verifying ' + filename)
+        hash_url = url + filename + '_md5.txt'
+        hash_filename = filename + '_md5.txt'
+        urllib.request.urlretrieve(hash_url, hash_filename)
+        filename = filename + '.tar.gz'
+        with open(hash_filename) as hash_file:
+            hashes = hash_file.read().splitlines()
+        os.remove(hash_filename)
+        for hash in hashes:
+            if not hash in hash_table:
+                print('  > Downloading Files...')
+                urllib.request.urlretrieve(url + filename, filename)
+                with tarfile.open(filename) as tar:
+                    tar.extractall(cache)
+                os.remove(filename)
+                break #continue with next map
 
     print('All Done!')
 except KeyboardInterrupt:
@@ -71,6 +104,10 @@ except KeyboardInterrupt:
         os.remove(filename)
     except OSError:
         pass
+    try:
+        os.remove(hash_filename)
+    except OSError:
+        pass
     print('\nYou cancelled the download.')
-    
+
 quit()
